Auto-disble page cache module on error
