+++ projects/plugins/boost/app/assets/src/js/features/page-cache/health/health.tsx
@@ -1,34 +1,51 @@
-import { Button, Notice } from '@automattic/jetpack-components';
+import { Notice } from '@automattic/jetpack-components';
 import getErrorData from './lib/get-error-data';
 import { __ } from '@wordpress/i18n';
-import { type PageCacheError } from '$lib/stores/page-cache';
-
+import { usePageCacheSetup, type PageCacheError } from '$lib/stores/page-cache';
+import { MutationNotice } from '$features/ui';
+import { useEffect } from 'react';
+import { useSingleModuleState } from '$features/module/lib/stores';
 type HealthProps = {
-	error: PageCacheError;
-	setupCache: () => void;
+	error?: PageCacheError;
+	setup?: ReturnType< typeof usePageCacheSetup >;
+	setError: ( error: PageCacheError ) => void;
 };
 
-const Health = ( { error, setupCache }: HealthProps ) => {
+const Health = ( { setup, error, setError }: HealthProps ) => {
+	const [ pageCache, setPageCache ] = useSingleModuleState( 'page_cache' );
 	// Was there a problem trying to setup cache?
 	const errorData = getErrorData( error );
-	if ( errorData ) {
-		return (
-			<Notice
-				level="warning"
-				hideCloseButton={ true }
-				title={ errorData.title }
-				actions={ [
-					<Button size="small" weight="regular" onClick={ setupCache } key="try-again">
-						{ __( 'Try again', 'jetpack-boost' ) }
-					</Button>,
-				] }
-			>
-				<p>{ errorData.message }</p>
-			</Notice>
-		);
-	}
-
-	return null;
+	useEffect( () => {
+		if ( error && pageCache?.active ) {
+			setPageCache( false );
+		}
+	} );
+	return (
+		<>
+			{ setup && (
+				<MutationNotice
+					mutationId="page-cache-setup"
+					isPending={ setup.isPending }
+					isError={ setup.isError }
+					isSuccess={ setup.isSuccess && ! error }
+					savingMessage={ __( 'Setting up cache…', 'jetpack-boost' ) }
+					errorMessage={ __( 'An error occurred while setting up cache.', 'jetpack-boost' ) }
+					successMessage={ __( 'Cache setup complete.', 'jetpack-boost' ) }
+				/>
+			) }
+			{ errorData && error && error.dismissed !== true && (
+				<Notice
+					level="warning"
+					title={ errorData.title }
+					onClose={ () => {
+						setError( { ...error, dismissed: true } );
+					} }
+				>
+					<p>{ errorData.message }</p>
+				</Notice>
+			) }
+		</>
+	);
 };
 
 export default Health;
