+++ projects/plugins/boost/app/assets/src/js/features/page-cache/health/lib/get-error-data.tsx
@@ -189,13 +189,10 @@ type FormattedError = {
 	title: string;
 	message: React.ReactNode;
 };
-function getErrorData( status?: PageCacheError ): FormattedError {
+function getErrorData( status?: PageCacheError ): false | FormattedError {
 	// Fall back to unknown error if no status
 	if ( ! status ) {
-		return {
-			title: __( 'Unknown error', 'jetpack-boost' ),
-			message: __( 'An unknown error occurred.', 'jetpack-boost' ),
-		};
+		return false;
 	}
 
 	// Try to find a message based on status code
