+++ projects/plugins/boost/app/assets/src/js/features/page-cache/page-cache.module.scss
@@ -100,7 +100,7 @@
 		margin-right: 1em;
 		float: left;
 	}
-	
+
 	.clearfix {
 		clear: both;
 	}
